make config typescript
